print prompt and signal

+++ libft/ft_bzero.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_bzero.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: vminomiy <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: hbuisser <hbuisser@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2020/01/21 17:32:35 by vminomiy          #+#    #+#             */
-/*   Updated: 2020/01/27 16:43:18 by vminomiy         ###   ########.fr       */
+/*   Created: 2019/10/15 10:57:35 by hbuisser          #+#    #+#             */
+/*   Updated: 2019/10/18 15:29:04 by hbuisser         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_bzero(void *s, unsigned long int n)
+void	ft_bzero(void *s, size_t n)
 {
 	ft_memset(s, 0, n);
 }
